Remove `--use-http2` flag
A major use case for this repo is to enable IAP for access to data in GCS buckets. However, [IAP is currently incompatible with http/2](https://cloud.google.com/iap/docs/enabling-cloud-run#known_limitations), which results in an infinite redirect loop. This change reverts to the default deployment without http/2 enabled.